refactor(#792): split stream_isolation_tests into base + multi-channel suite (sweep)
stream_isolation_tests.rs 630 -> 412, + stream_isolation_tests_more.rs 227 (both
under cap; off debt). The multi-channel / split-mono / dual-mono isolation tests
move to the sibling; the binding fixtures (in_ep/out_ep/binding/bound_chain) become
pub(super) and the new file imports them via super::stream_isolation. The pinned
#[ignore] (#350-vs-#400 documented conflict) moves verbatim. engine tests (36
groups) pass, zero warnings.
